removed some warnings

+++ src/main/java/one/ryd/insider/models/session/aggregation/SessionAlikeDto.java
@@ -27,8 +27,8 @@ public final class SessionAlikeDto {
   public Double confidence;
 
   @JsonIgnore
-  public Integer hits;
+  public Long hits;
 
   @JsonIgnore
-  public Integer intersection;
+  public Double intersection;
 }
